prototype CaseMap objects
It would be convenient to be able to hand off casemaps to Channel objects, but
pretty memory-intensive to store a string for each one. Instead provide a
CaseMap flyweight that caches little stringifying scalar-type objects for each
casemap; objects can hang on to a reference to a ValidCaseMapObject coercible
from a ValidCaseMap instead of a string. Hopefully slightly cheaper.

	new file:   lib/IRC/Server/Toolkit/CaseMap.pm
	modified:   lib/IRC/Server/Toolkit/Collection/Users.pm
	modified:   lib/IRC/Server/Toolkit/Types.pm
	modified:   t/types.t